+++ src/words/antonym.js
@@ -1,9 +1,9 @@
-const Result = require('./httpsclient.js');
+const Result = require('./httpsclient');
 
-async function getSynonym (findword) {
+async function getAntonym (findword){
     var request ="word/"+String(findword)+"/relatedWords";
     console.log(request);
     await Result.getResult(request);
 }
 
-module.exports.getSynonym = getSynonym;
+module.exports.getAntonym = getAntonym;
