+++ src/words/fulldetails.js
@@ -4,7 +4,6 @@ const antonym = require('./antonym');
 const example = require('./examples');
 
 async function getFulldetails(findword) {
-    //var request ="word/"+String(findword)+"/definitions";
     await definition.getDefinition(findword);
     await synonym.getSynonym(findword);
     await antonym.getAntonym(findword);
