+++ src/words/randomn.js
@@ -1,0 +1,9 @@
+const Result = require('./httpsclient');
+
+async function getRandomn() {
+    var request ="words/randomWord";
+    console.log(request);
+    await Result.getResult(request);
+}
+
+module.exports.getRandomn = getRandomn;
